Refactor service package name
