+++ src/main/java/se/iths/webservices/tonahs/postservice/service/PostService.java
@@ -1,4 +1,4 @@
-package se.iths.webservices.tonahs.postservice.services;
+package se.iths.webservices.tonahs.postservice.service;
 
 import se.iths.webservices.tonahs.postservice.dto.PostDto;
 import se.iths.webservices.tonahs.postservice.dto.PostNewDto;
